Adds comments

+++ src/main/java/com/benitosaell/client/controller/MovieController.java
@@ -6,6 +6,7 @@ import java.util.LinkedList;
 import java.util.List;
 
 import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpSession;
 
 import org.springframework.beans.propertyeditors.CustomDateEditor;
 import org.springframework.http.ResponseEntity;
@@ -13,7 +14,6 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.WebDataBinder;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.InitBinder;
 import org.springframework.web.bind.annotation.ModelAttribute;
@@ -31,10 +31,21 @@ import com.benitosaell.client.util.Util;
 @Controller
 @RequestMapping("/peliculas")
 public class MovieController {
+	//Variable para las peticiones a la api.
 	RestTemplate restTemplate = new RestTemplate();
-
+	
+	/*
+	 * Interfaz para la lista de peliculas registradas en la base de datos.
+	 * 
+	 * @param model: Cargar lista de peliculas para la vista.
+	 * @param sessionMain: Restringir el acceso.
+	 * @return : Vista de la lista de peliculas.
+	*/
 	@GetMapping("/")
-	private String getMovies(Model model) {
+	private String getMovies(Model model, HttpSession sessionMain ) {
+		if(sessionMain.getAttribute("userAdmin")==null) {
+			return "redirect:/";
+		}
 		List<Movie> movies = new LinkedList<>();
 		ResponseEntity<Movie[]> response = restTemplate.getForEntity("http://localhost:3000/api/peliculas/peliculas",
 				Movie[].class);
@@ -46,13 +57,34 @@ public class MovieController {
 		return "movies/ListMovies";
 	}
 
+	/*
+	 * Carga el formulario para crear una nueva pelicula.
+	 * 
+	 * @param movie: Enlace con la información de la pelicula.
+	 * @param sessionMain: Restringir el acceso.
+	 * @return : Vista del formulario.
+	*/
 	@GetMapping("/nueva")
-	private String newMovie(@ModelAttribute Movie movie, Model model) {
+	private String newMovie(@ModelAttribute Movie movie, HttpSession sessionMain) {
+		if(sessionMain.getAttribute("userAdmin")==null) {
+			return "redirect:/";
+		}
 		return "movies/Movie";
 	}
 
+	/*
+	 * Carga el formulario editar pelicula.
+	 * 
+	 * @param id: De la pelicula a editar.
+	 * @param model: Información del la pelicula a editar.
+	 * @param sessionMain: Restringir el acceso.
+	 * @return : Vista del formulario.
+	*/
 	@GetMapping("/editar/{id}")
-	public String edit(@PathVariable("id") int idMovie, Model model) {
+	public String edit(@PathVariable("id") int idMovie, Model model, HttpSession sessionMain) {
+		if(sessionMain.getAttribute("userAdmin")==null) {
+			return "redirect:/";
+		}
 		Movie movie = new Movie();
 		ResponseEntity<Movie> response = restTemplate.getForEntity("http://localhost:3000/api/peliculas/ver/" + idMovie,
 				Movie.class);
@@ -63,8 +95,19 @@ public class MovieController {
 		return "movies/Movie";
 	}
 
+	/*
+	 * Eliminar pelicula.
+	 * 
+	 * @param id: Referencia de la pelicula.
+	 * @param attribute: Para crear mensajes.
+	 * @param sessionMain: Restringir el acceso.
+	 * @return : Vista del formulario.
+	*/
 	@GetMapping("/eliminar/{id}")
-	public String delete(@PathVariable("id") int idMovie, RedirectAttributes attribute) {
+	public String delete(@PathVariable("id") int idMovie, HttpSession sessionMain, RedirectAttributes attribute) {
+		if(sessionMain.getAttribute("userAdmin")==null) {
+			return "redirect:/";
+		}
 		System.out.println("IDMovie: "+idMovie);
 		restTemplate.delete("http://localhost:3000/api/peliculas/eliminar/"+idMovie);
 		restTemplate.delete("http://localhost:3000/api/comentarios/comentarios/"+idMovie);
@@ -72,6 +115,16 @@ public class MovieController {
 		return "redirect:/peliculas/";
 	}
 
+	/*
+	 * Guardar información de la pelicula.
+	 * 
+	 * @param movie: Referencia de la pelicula.
+	 * @param result: Verificar errores.
+	 * @param attribute: Para crear mensajes.
+	 * @param multiPart: Imagen.
+	 * @param request.
+	 * @return : Vista del formulario.
+	*/
 	@PostMapping(value = "/guardar")
 	public String save(@ModelAttribute Movie movie, BindingResult result, RedirectAttributes attributes,
 			@RequestParam("fileImage") MultipartFile multiPart, HttpServletRequest request) {
@@ -83,28 +136,25 @@ public class MovieController {
 			}
 
 			if (result.hasErrors()) {
-				System.out.println("Existieron errores");
 				return "movies/Movie";
 			}
-
-			System.out.println("IdClient: " + movie);
-
 			ResponseEntity<Movie> response =
 			restTemplate.postForEntity("http://localhost:3000/api/peliculas/crear",
 			movie,Movie.class);
-			System.out.println("Respuesta " + response.getBody());
 			movie = response.getBody();
 
 			attributes.addFlashAttribute("message", "La pelicula fue insertada correctamente");
 			return "redirect:/peliculas/";
 			
 		}catch(Exception err) {
-			System.out.println("Error: "+err);
 			return "movies/Movie";
 		}
 		
 	}
 
+	/*
+	 * Formato a la fecha.
+	*/
 	@InitBinder
 	public void initBinder(WebDataBinder binder) {
 		SimpleDateFormat dateFormat = new SimpleDateFormat("dd-MM-yyyy");
